// src/api/generated/projects.d.ts
export type AssetId = number;
export type AssetType = "image" | "video";
export type Url = string;
export type CoverAssetId = number | null;
export type CreatedAt = string;
export type EpisodeCount = number;
export type Id = number;
export type Name1 = string;
export type Status = number;
export type UpdatedAt = string;
export type CoverAssetId1 = number | null;
export type CreatedAt1 = string;
export type CreatorId = number;
export type EpisodeNo = number;
export type Id1 = number;
export type Name2 = string;
export type ProjectId = number;
export type UpdatedAt1 = string;
// ...
export type Items1 = EpisodeView[];
export type Page3 = number;
export type PageSize3 = number;
export type Total1 = number;
// ...
export interface CoverView {
  asset_id: AssetId;
  asset_type: AssetType;
  url: Url;
  [k: string]: unknown;
}
export interface ProjectCreateResponse {
  default_episode: EpisodeView;
  project: ProjectView;
  [k: string]: unknown;
}
export interface EpisodeView {
  cover?: CoverView | null;
  cover_asset_id?: CoverAssetId1;
  created_at: CreatedAt1;
  creator_id: CreatorId;
  episode_no: EpisodeNo;
  id: Id1;
  name: Name2;
  project_id: ProjectId;
  updated_at: UpdatedAt1;
  [k: string]: unknown;
}
// ...
export interface EpisodeListResponse {
  items: Items1;
  page?: Page3;
  page_size?: PageSize3;
  total?: Total1;
  [k: string]: unknown;
}